DSDEGP-1639: code review feedback
- merge UpsertId and UpsertIdGenerator
- restrict access to UpsertId creation
- update comment
